[AOT] Move LLVM-related code into separate directory
* Rename cmake config WASMEDGE_BUILD_AOT_RUNTIME into WASMEDGE_USE_LLVM
  * Add a warning if old config was supplied
* Separate out `CodeGen` function from `Compiler` class
* Add `LLVM::Data` class for transfer `LLVM::Context` and `LLVM::Module`
* Use `fpexcept.strict` for correct NAN-related behaviors
