+++ examples/embed_cxx/CMakeLists.txt
@@ -5,7 +5,7 @@ project(embed_cxx CXX)
 set(CMAKE_CXX_STANDARD 17)
 
 set(WASMEDGE_SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/../..)
-set(WASMEDGE_BUILD_AOT_RUNTIME OFF)
+set(WASMEDGE_USE_LLVM OFF)
 if (CMAKE_GENERATOR STREQUAL Ninja)
   set(CMAKE_JOB_POOLS "link=2")
   set(CMAKE_JOB_POOL_LINK link)
