+++ lib/CMakeLists.txt
@@ -1,8 +1,9 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: 2019-2022 Second State INC
 
-if(WASMEDGE_BUILD_AOT_RUNTIME)
+if(WASMEDGE_USE_LLVM)
   add_subdirectory(aot)
+  add_subdirectory(llvm)
 endif()
 add_subdirectory(common)
 add_subdirectory(system)
