+++ lib/api/CMakeLists.txt
@@ -3,8 +3,8 @@
 
 find_package(spdlog QUIET)
 add_definitions(-DWASMEDGE_COMPILE_LIBRARY)
-if(WASMEDGE_BUILD_AOT_RUNTIME)
-  add_definitions(-DWASMEDGE_BUILD_AOT_RUNTIME)
+if(WASMEDGE_USE_LLVM)
+  add_definitions(-DWASMEDGE_USE_LLVM)
 endif()
 if(WASMEDGE_BUILD_WASI_NN_RPC)
   add_definitions(-DWASMEDGE_BUILD_WASI_NN_RPC)
@@ -82,10 +82,11 @@ target_link_libraries(wasmedgeCAPI
   wasmedgeDriver
 )
 
-if (WASMEDGE_BUILD_AOT_RUNTIME)
+if(WASMEDGE_USE_LLVM)
   target_link_libraries(wasmedgeCAPI
     PUBLIC
     wasmedgeAOT
+    wasmedgeLLVM
   )
 endif()
 
@@ -155,12 +156,13 @@ if(WASMEDGE_BUILD_STATIC_LIB)
   wasmedge_add_static_lib_component_command(wasmedgeVM)
   wasmedge_add_static_lib_component_command(wasmedgeDriver)
 
-  if(WASMEDGE_BUILD_AOT_RUNTIME)
+  if(WASMEDGE_USE_LLVM)
     foreach(LIB_NAME IN LISTS WASMEDGE_LLVM_LINK_STATIC_COMPONENTS)
       wasmedge_add_libs_component_command(${LIB_NAME})
     endforeach()
     wasmedge_add_static_lib_component_command(utilBlake3)
     wasmedge_add_static_lib_component_command(wasmedgeAOT)
+    wasmedge_add_static_lib_component_command(wasmedgeLLVM)
   endif()
 
   if(CMAKE_AR_NAME STREQUAL "libtool")
