+++ lib/driver/CMakeLists.txt
@@ -24,7 +24,7 @@ if(WASMEDGE_BUILD_WASI_NN_RPC)
   )
 endif()
 
-if(WASMEDGE_BUILD_AOT_RUNTIME)
+if(WASMEDGE_USE_LLVM)
   target_link_libraries(wasmedgeDriver
     PRIVATE
     wasmedgeLoader
@@ -32,6 +32,7 @@ if(WASMEDGE_BUILD_AOT_RUNTIME)
     wasmedgePO
     wasmedgeVM
     wasmedgeAOT
+    wasmedgeLLVM
   )
 else()
   target_link_libraries(wasmedgeDriver
@@ -43,9 +44,9 @@ else()
   )
 endif()
 
-if(WASMEDGE_BUILD_AOT_RUNTIME)
+if(WASMEDGE_USE_LLVM)
   target_compile_definitions(wasmedgeDriver
     PRIVATE
-    -DWASMEDGE_BUILD_AOT_RUNTIME
+    -DWASMEDGE_USE_LLVM
   )
 endif()
