+++ test/CMakeLists.txt
@@ -60,8 +60,9 @@ else()
   set(GTEST_BOTH_LIBRARIES "gtest;gtest_main")
 endif()
 
-if (WASMEDGE_BUILD_AOT_RUNTIME)
+if (WASMEDGE_USE_LLVM)
   add_subdirectory(aot)
+  add_subdirectory(llvm)
   add_subdirectory(mixcall)
 endif()
 add_subdirectory(common)
