+++ test/llvm/CMakeLists.txt
@@ -1,0 +1,18 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: 2019-2022 Second State INC
+
+wasmedge_add_executable(wasmedgeLLVMCoreTests
+  LLVMcoreTest.cpp
+)
+
+add_test(wasmedgeLLVMCoreTests wasmedgeLLVMCoreTests)
+
+target_link_libraries(wasmedgeLLVMCoreTests
+  PRIVATE
+  std::filesystem
+  ${GTEST_BOTH_LIBRARIES}
+  wasmedgeTestSpec
+  wasmedgeLoader
+  wasmedgeLLVM
+  wasmedgeVM
+)
