+++ test/llvm/LLVMcoreTest.cpp
@@ -14,10 +14,11 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include "aot/compiler.h"
 #include "common/defines.h"
 #include "common/log.h"
 #include "vm/vm.h"
+#include "llvm/codegen.h"
+#include "llvm/compiler.h"
 
 #include "../spec/hostfunc.h"
 #include "../spec/spectest.h"
@@ -60,24 +61,27 @@ TEST_P(NativeCoreTest, TestSuites) {
     CopyConf.getCompilerConfigure().setOptimizationLevel(
         WasmEdge::CompilerConfigure::OptimizationLevel::O0);
     CopyConf.getCompilerConfigure().setDumpIR(true);
-    WasmEdge::AOT::Compiler Compiler(CopyConf);
+    WasmEdge::LLVM::Compiler Compiler(CopyConf);
+    WasmEdge::LLVM::CodeGen CodeGen(CopyConf);
     auto Path = std::filesystem::u8path(Filename);
     Path.replace_extension(std::filesystem::u8path(WASMEDGE_LIB_EXTENSION));
     const auto SOPath = Path.u8string();
-    auto Data = *Loader.loadFile(Filename);
+    std::vector<WasmEdge::Byte> Data;
     std::unique_ptr<WasmEdge::AST::Module> Module;
-    if (auto Res = Loader.parseModule(Data)) {
-      Module = std::move(*Res);
-    } else {
-      return Unexpect(Res);
-    }
-    if (auto Res = ValidatorEngine.validate(*Module); !Res) {
-      return Unexpect(Res);
-    }
-    if (auto Res = Compiler.compile(Data, *Module, SOPath); !Res) {
-      return Unexpect(Res);
-    }
-    return SOPath;
+    return Loader.loadFile(Filename)
+        .and_then([&](auto Result) noexcept {
+          Data = std::move(Result);
+          return Loader.parseModule(Data);
+        })
+        .and_then([&](auto Result) noexcept {
+          Module = std::move(Result);
+          return ValidatorEngine.validate(*Module);
+        })
+        .and_then([&]() noexcept { return Compiler.compile(*Module); })
+        .and_then([&](auto Result) noexcept {
+          return CodeGen.codegen(Data, std::move(Result), SOPath);
+        })
+        .and_then([&]() noexcept { return Expect<std::string>{SOPath}; });
   };
   T.onModule = [&VM, &Compile](const std::string &ModName,
                                const std::string &Filename) -> Expect<void> {
@@ -166,24 +170,27 @@ TEST_P(CustomWasmCoreTest, TestSuites) {
     CopyConf.getCompilerConfigure().setOptimizationLevel(
         WasmEdge::CompilerConfigure::OptimizationLevel::O0);
     CopyConf.getCompilerConfigure().setDumpIR(true);
-    WasmEdge::AOT::Compiler Compiler(CopyConf);
+    WasmEdge::LLVM::Compiler Compiler(CopyConf);
+    WasmEdge::LLVM::CodeGen CodeGen(CopyConf);
     auto Path = std::filesystem::u8path(Filename);
     Path.replace_extension(std::filesystem::u8path(".aot.wasm"));
     const auto SOPath = Path.u8string();
-    auto Data = *Loader.loadFile(Filename);
+    std::vector<WasmEdge::Byte> Data;
     std::unique_ptr<WasmEdge::AST::Module> Module;
-    if (auto Res = Loader.parseModule(Data)) {
-      Module = std::move(*Res);
-    } else {
-      return Unexpect(Res);
-    }
-    if (auto Res = ValidatorEngine.validate(*Module); !Res) {
-      return Unexpect(Res);
-    }
-    if (auto Res = Compiler.compile(Data, *Module, SOPath); !Res) {
-      return Unexpect(Res);
-    }
-    return SOPath;
+    return Loader.loadFile(Filename)
+        .and_then([&](auto Result) noexcept {
+          Data = std::move(Result);
+          return Loader.parseModule(Data);
+        })
+        .and_then([&](auto Result) noexcept {
+          Module = std::move(Result);
+          return ValidatorEngine.validate(*Module);
+        })
+        .and_then([&]() noexcept { return Compiler.compile(*Module); })
+        .and_then([&](auto Result) noexcept {
+          return CodeGen.codegen(Data, std::move(Result), SOPath);
+        })
+        .and_then([&]() noexcept { return Expect<std::string>{SOPath}; });
   };
   T.onModule = [&VM, &Compile](const std::string &ModName,
                                const std::string &Filename) -> Expect<void> {
@@ -282,12 +289,15 @@ TEST(AsyncRunWsmFile, NativeInterruptTest) {
   WasmEdge::VM::VM VM(Conf);
   WasmEdge::Loader::Loader Loader(Conf);
   WasmEdge::Validator::Validator ValidatorEngine(Conf);
-  WasmEdge::AOT::Compiler Compiler(Conf);
+  WasmEdge::LLVM::Compiler Compiler(Conf);
+  WasmEdge::LLVM::CodeGen CodeGen(Conf);
   auto Path = std::filesystem::temp_directory_path() /
               std::filesystem::u8path("AOTcoreTest" WASMEDGE_LIB_EXTENSION);
   auto Module = *Loader.parseModule(AsyncWasm);
   ASSERT_TRUE(ValidatorEngine.validate(*Module));
-  ASSERT_TRUE(Compiler.compile(AsyncWasm, *Module, Path));
+  auto Data = Compiler.compile(*Module);
+  ASSERT_TRUE(Data);
+  ASSERT_TRUE(CodeGen.codegen(AsyncWasm, std::move(*Data), Path));
   {
     auto Timeout =
         std::chrono::system_clock::now() + std::chrono::milliseconds(1);
@@ -320,12 +330,15 @@ TEST(AsyncExecute, NativeInterruptTest) {
   WasmEdge::VM::VM VM(Conf);
   WasmEdge::Loader::Loader Loader(Conf);
   WasmEdge::Validator::Validator ValidatorEngine(Conf);
-  WasmEdge::AOT::Compiler Compiler(Conf);
+  WasmEdge::LLVM::Compiler Compiler(Conf);
+  WasmEdge::LLVM::CodeGen CodeGen(Conf);
   auto Path = std::filesystem::temp_directory_path() /
               std::filesystem::u8path("AOTcoreTest" WASMEDGE_LIB_EXTENSION);
   auto Module = *Loader.parseModule(AsyncWasm);
   ASSERT_TRUE(ValidatorEngine.validate(*Module));
-  ASSERT_TRUE(Compiler.compile(AsyncWasm, *Module, Path));
+  auto Data = Compiler.compile(*Module);
+  ASSERT_TRUE(Data);
+  ASSERT_TRUE(CodeGen.codegen(AsyncWasm, std::move(*Data), Path));
   ASSERT_TRUE(VM.loadWasm(Path));
   ASSERT_TRUE(VM.validate());
   ASSERT_TRUE(VM.instantiate());
@@ -361,12 +374,15 @@ TEST(AsyncRunWsmFile, CustomWasmInterruptTest) {
   WasmEdge::VM::VM VM(Conf);
   WasmEdge::Loader::Loader Loader(Conf);
   WasmEdge::Validator::Validator ValidatorEngine(Conf);
-  WasmEdge::AOT::Compiler Compiler(Conf);
+  WasmEdge::LLVM::Compiler Compiler(Conf);
+  WasmEdge::LLVM::CodeGen CodeGen(Conf);
   auto Path = std::filesystem::temp_directory_path() /
               std::filesystem::u8path("AOTcoreTest.aot.wasm");
   auto Module = *Loader.parseModule(AsyncWasm);
   ASSERT_TRUE(ValidatorEngine.validate(*Module));
-  ASSERT_TRUE(Compiler.compile(AsyncWasm, *Module, Path));
+  auto Data = Compiler.compile(*Module);
+  ASSERT_TRUE(Data);
+  ASSERT_TRUE(CodeGen.codegen(AsyncWasm, std::move(*Data), Path));
   {
     auto Timeout =
         std::chrono::system_clock::now() + std::chrono::milliseconds(1);
@@ -399,12 +415,15 @@ TEST(AsyncExecute, CustomWasmInterruptTest) {
   WasmEdge::VM::VM VM(Conf);
   WasmEdge::Loader::Loader Loader(Conf);
   WasmEdge::Validator::Validator ValidatorEngine(Conf);
-  WasmEdge::AOT::Compiler Compiler(Conf);
+  WasmEdge::LLVM::Compiler Compiler(Conf);
+  WasmEdge::LLVM::CodeGen CodeGen(Conf);
   auto Path = std::filesystem::temp_directory_path() /
               std::filesystem::u8path("AOTcoreTest.aot.wasm");
   auto Module = *Loader.parseModule(AsyncWasm);
   ASSERT_TRUE(ValidatorEngine.validate(*Module));
-  ASSERT_TRUE(Compiler.compile(AsyncWasm, *Module, Path));
+  auto Data = Compiler.compile(*Module);
+  ASSERT_TRUE(Data);
+  ASSERT_TRUE(CodeGen.codegen(AsyncWasm, std::move(*Data), Path));
   ASSERT_TRUE(VM.loadWasm(Path));
   ASSERT_TRUE(VM.validate());
   ASSERT_TRUE(VM.instantiate());
