+++ utils/docker/Dockerfile.debian-static
@@ -75,14 +75,14 @@ RUN cmake -S /src -B /build -G Ninja \
     -DCMAKE_INSTALL_PREFIX=/install \
     -DWASMEDGE_BUILD_PACKAGE="TGZ" \
     -DWASMEDGE_BUILD_TESTS=OFF \
-    -DWASMEDGE_BUILD_AOT_RUNTIME=ON \
     -DWASMEDGE_BUILD_SHARED_LIB=OFF \
     -DWASMEDGE_BUILD_STATIC_LIB=ON \
     -DWASMEDGE_BUILD_TOOLS=OFF \
     -DWASMEDGE_BUILD_PLUGINS=OFF \
     -DWASMEDGE_BUILD_EXAMPLE=OFF \
     -DWASMEDGE_LINK_LLVM_STATIC=ON \
-    -DWASMEDGE_LINK_TOOLS_STATIC=ON
+    -DWASMEDGE_LINK_TOOLS_STATIC=ON \
+    -DWASMEDGE_USE_LLVM=ON
 
 RUN cmake --build /build -- install
 RUN cmake --build /build -- package
